feat(ui): unify UI layout for manage client task and add new task
