+++ src/app/management_dashboard/add_care_items/page.tsx
